Update and rename pandrator_start_minimal_xtts.py to pandrator_installer_launcher_xtts.py
The XTTS installer has received a GUI; it's possible to install XTTS with or without GPU support and to choose weather to run XTTS with or without --deepspeed and --lowvram arguments. It also functions as a launcher. In the future Silero, VoiceCraft, RVC etc. will be added to it as well.